batches/query-generator/requirements.txt
# AWS SDK and services
boto3>=1.34.0
botocore>=1.34.0

# Database connectivity
psycopg[binary]>=3.1.0

# Google AI
google-genai>=0.1.0

# Core Python libraries (usually included but specified for clarity)
requests>=2.31.0
urllib3>=2.0.0

# JSON and data handling
python-dateutil>=2.8.0

# Logging and monitoring (optional but recommended for batch jobs)
structlog>=23.1.0